fixes for auto generation random uuids on create.
