+++ package.json
@@ -7,12 +7,13 @@
   "author": "Your Name Joseph Atzamis iwsifa@gmail.com",
   "main": "src/index.ts",
   "scripts": {
-    "dev": "bun run --hot src/server.ts",
+    "dev": "bun run --hot src/index.ts",
     "build": "bun run ./build.ts",
     "db:generate": "drizzle-kit generate",
     "db:migrate": "bun run src/db/database/migrate.ts",
     "db:drop": "drizzle-kit drop",
     "db:pull": "drizzle-kit introspect",
+    "db:push": "drizzle-kit push",
     "db:studio": "drizzle-kit studio",
     "db:check": "drizzle-kit check",
     "lint": "eslint .",
@@ -20,11 +21,10 @@
   },
   "dependencies": {
     "@hono/node-server": "^1.13.2",
-    "@hono/swagger-ui": "^0.4.1",
     "@hono/zod-openapi": "^0.16.4",
-    "@libsql/client": "^0.14.0",
     "@scalar/hono-api-reference": "^0.5.156",
     "@supabase/supabase-js": "^2.45.6",
+    "@types/uuid": "^10.0.0",
     "dotenv": "^16.4.5",
     "dotenv-expand": "^11.0.6",
     "drizzle-orm": "^0.35.2",
@@ -34,11 +34,11 @@
     "pino": "^9.4.0",
     "pino-pretty": "^11.2.2",
     "postgres": "^3.4.4",
+    "uuid": "^10.0.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^3.8.0",
-    "@biomejs/biome": "^1.9.4",
     "@types/bun": "latest",
     "drizzle-kit": "^0.26.2",
     "esbuild": "^0.24.0",
