+++ src/db/database/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from "./users";
+export * from "./wishes";
